Extracted the bond settings to its own component
